Add @tangleid/mnid package
To make the package can be easily maintained, move the MNID functions
to "@tangleid/mnid" the package, and convert it to TypeScript project.

Since TypeScript compiler does not clean up the output directory
before compiling, add script to up output directory and run it before
building the packages.

Since the README documentation cannot be generated from the TypeScript
files directly, compile these files to the JavaScript, then generate
documentation from them.
