+++ package.json
@@ -6,9 +6,10 @@
   ],
   "scripts": {
     "init": "yarn install && lerna bootstrap",
-    "test": "lerna exec -- yarn run test",
-    "build": "lerna exec -- yarn run build",
-    "docs": "lerna exec -- yarn run docs",
+    "test": "yarn run build && lerna exec -- yarn run test",
+    "docs": "yarn run build && lerna exec -- yarn run docs",
+    "clean": "lerna exec --parallel -- rimraf lib",
+    "build": "yarn run clean && lerna exec -- yarn run build",
     "audit": "yarn audit && lerna exec -- yarn audit"
   },
   "devDependencies": {
@@ -17,6 +18,8 @@
     "@babel/plugin-proposal-class-properties": "^7.4.0",
     "@babel/plugin-transform-runtime": "^7.1.0",
     "@babel/preset-env": "^7.1.0",
+    "@types/jest": "^23.3.13",
+    "@types/node": "^11.11.4",
     "babel-core": "^7.0.0-bridge.0",
     "dmd-clear": "^0.1.2",
     "husky": "^1.3.1",
@@ -24,7 +27,10 @@
     "jsdoc-to-markdown": "^4.0.1",
     "lerna": "^3.13.1",
     "lint-staged": "^7.3.0",
-    "prettier": "^1.14.3"
+    "prettier": "^1.14.3",
+    "rimraf": "^2.6.3",
+    "ts-jest": "^23.10.5",
+    "typescript": "^3.2.4"
   },
   "husky": {
     "hooks": {
