+++ packages/mnid/package.json
@@ -1,0 +1,26 @@
+{
+  "name": "@tangleid/mnid",
+  "version": "1.1.1",
+  "description": "Utilities for generating Multi Network Identifier (MNID)",
+  "author": "TangleID Developers",
+  "main": "lib/index.js",
+  "typings": "typings",
+  "scripts": {
+    "test": "jest",
+    "build": "tsc",
+    "docs": "jsdoc2md --no-cache --plugin dmd-clear -t README.template --files './lib/**/*.js' > README.md"
+  },
+  "license": "MIT",
+  "repository": {
+    "type": "git",
+    "url": "https://github.com/TangleID/tangleid-core/tree/master/packages/did"
+  },
+  "dependencies": {
+    "base-x": "^3.0.4",
+    "buffer": "^5.2.0",
+    "js-sha3": "^0.8.0"
+  },
+  "publishConfig": {
+    "access": "public"
+  }
+}
