+++ packages/mnid/src/types.ts
@@ -1,0 +1,7 @@
+export type Payload = Buffer[];
+export type Mnid = string;
+
+export type MnidModel = {
+  network: string;
+  address: string;
+};
