+++ packages/mnid/tsconfig.json
@@ -1,0 +1,13 @@
+{
+  "compilerOptions": {
+    "types": ["node", "jest"],
+    "lib": ["dom", "es2015"],
+    "noImplicitThis": true,
+    "strict": true,
+    "outDir": "lib",
+    "declaration": true,
+    "declarationDir": "typings",
+    "sourceMap": true
+  },
+  "include": ["src"]
+}
